refactor: folder and file structure
