+++ apps/api/v2/src/modules/auth/decorator/index.ts
@@ -1,1 +1,2 @@
-export * from "./get-user.decorator";
+export * from "./get-user/get-user.decorator";
+export * from "./roles/roles.decorator";
