+++ apps/api/v2/src/modules/auth/guard/index.ts
@@ -1,1 +1,3 @@
-export * from "./next-auth";
+export * from "./next-auth/next-auth.guard";
+export * from "./organization/organization.guard";
+export * from "./roles/roles.guard";
