+++ apps/api/v2/src/modules/auth/guard/next-auth/next-auth.guard.ts
