+++ apps/api/v2/src/modules/auth/strategy/api-key-auth/api-key-auth.strategy.ts
