+++ apps/api/v2/src/modules/auth/strategy/index.ts
@@ -1,1 +1,2 @@
-export * from "./next-auth";
+export * from "./next-auth/next-auth.strategy";
+export * from "./api-key-auth/api-key-auth.strategy";
