+++ apps/api/v2/src/modules/auth/strategy/next-auth/next-auth.strategy.ts
